add vel gamma and get rid of some warnings

+++ Source/PolygonDisplay.h
@@ -38,7 +38,7 @@ public:
 
     void stopDrawing() { oscilloscope2D->stop(); }
     void startDrawing() { oscilloscope2D->start(); }
-    void updateBlocksize(float bs) { oscilloscope2D->updateBlocksize(bs); }
+    void updateBlocksize(int bs) { oscilloscope2D->updateBlocksize(bs); }
 
 private:
     AudioProcessorValueTreeState& parameters;
